final code all done

+++ Assets/C#/Titli/GamePlay/Titli_Timer.cs
@@ -357,6 +357,6 @@ waitForBetScreen.SetActive(false);
         public string balance;
         public int gametimer;
         public int userDailyWin;
-        public int currentWin;
+        public double currentWin;
     }
 }
